requirements.txt
# =========================
# Core Python
# =========================
numpy>=1.23
pandas>=1.5
scipy>=1.10

# =========================
# Data & Finance
# =========================
yfinance>=0.2.28
ta>=0.10.2

# =========================
# Machine Learning
# =========================
scikit-learn>=1.2
joblib>=1.3

# =========================
# Deep Learning
# =========================
torch>=2.0
torchvision>=0.15

# =========================
# Reinforcement Learning
# =========================
stable-baselines3>=2.1
gymnasium>=0.29

# =========================
# Regime Detection
# =========================
hmmlearn>=0.3
statsmodels>=0.14

# =========================
# NLP / News Sentiment
# =========================
transformers>=4.35
torchtext>=0.15
vaderSentiment>=3.3

# =========================
# Visualization
# =========================
matplotlib>=3.7
plotly>=5.18

# =========================
# Web App
# =========================
streamlit>=1.29

# =========================
# Utilities
# =========================
tqdm>=4.66
python-dotenv>=1.0
nsepython>=1.0
